+++ Frameworks/CloudrailSI.framework/Headers/CRFacebookPage.h
@@ -1,0 +1,13 @@
+
+#import <Foundation/Foundation.h>
+#import "CRSocialProtocol.h"
+#import "CRAuthenticatingProtocol.h"
+
+@interface CRFacebookPage : NSObject <CRSocialProtocol,CRAuthenticatingProtocol>
+@property (weak, nonatomic) id target;
+
+-(instancetype)initWithPageName:(NSString *)pageName clientID:(NSString *)clientID clientSecret:(NSString *)clientSecret;
+
+
+
+@end