+++ Frameworks/CloudrailSI.framework/Headers/CRGoogleCloudPlatform.h
@@ -1,0 +1,12 @@
+
+#import <Foundation/Foundation.h>
+#import "CRBusinessCloudStorageProtocol.h"
+
+@interface CRGoogleCloudPlatform : NSObject <CRBusinessCloudStorageProtocol>
+@property (weak, nonatomic) id target;
+
+-(instancetype)initWithClientEmail:(NSString *)clientEmail privateKey:(NSString *)privateKey projectId:(NSString *)projectId;
+
+
+
+@end