+++ Frameworks/CloudrailSI.framework/Headers/CRTwitter.h
@@ -6,9 +6,9 @@
 @interface CRTwitter : NSObject <CRProfileProtocol, CRSocialProtocol>
 @property (weak, nonatomic) id target;
 
-- (instancetype)initWithClientId:(NSString *)clientID clientSecret:(NSString *)clientSecret redirectUri:(NSString *)redirectUri;
+-(instancetype)initWithClientId:(NSString *)clientID clientSecret:(NSString *)clientSecret redirectUri:(NSString *)redirectUri;
 
 //Specific for CloudStorage Interface
-- (instancetype)initWithClientId:(NSString *)clientId clientSecret: (NSString *)clientSecret;
+-(instancetype)initWithClientId:(NSString *)clientId clientSecret: (NSString *)clientSecret;
 
 @end